core/requirements.txt
asgiref==3.11.1
attrs==26.1.0
autobahn==24.4.2
Automat==25.4.16
boto3==1.42.71
botocore==1.42.73
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.6
constantly==23.10.4
cryptography==46.0.5
daphne==4.2.1
Django==4.2.29
django-cors-headers==4.9.0
django-storages==1.14.2
djangorestframework==3.16.1
drf-yasg==1.21.15
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
jmespath==1.1.0
packaging==26.0
pillow==11.3.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==26.0.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
reportlab==4.4.10
requests==2.32.5
s3transfer==0.16.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
tomli==2.4.0
Twisted==25.5.0
txaio==23.6.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==1.26.20
zope.interface==8.0.1
whitenoise==6.11.0
